docs: readme atualizado

+++ 03 - Conversão de altura/ConversaoDeAlturaB.java
@@ -15,7 +15,7 @@ public class ConversaoDeAlturaB {
 
             double polegadas = Math.round(sobraEmPes * 12);
 
-            System.out.printf("Altura pés e polegadas: %.2f %.2f%n" , pes, polegadas);
+            System.out.printf("Altura pés e polegadas: %d %d%n" , (int) pes, (int) polegadas);
             scanner.close();
         }
     }
